templates/backend-fastapi/_supabase/pyproject.toml
[project]
name = "__PROJECT_KEBAB__-backend"
version = "0.1.0"
description = "__DESCRIPTION__"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.2",
  "python-multipart>=0.0.12",
  "supabase>=2.9.0",
  "PyJWT>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
